Add `tox` as test runner + update Travis configuration.
The new test runner is needed as Travis does not know `python2` and `python3`.
(BTW I have the same problem on my computer, so I cannot use the previous
approach with the Makefile without modifications on my system.)

The tests in grammar/test_grammar.py are not run because I was not able to find
out how to run `make` in the `grammar` directory before installing the package.

See https://tox.readthedocs.io for documentation on tox and
https://docs.pytest.org for documentation on pytest. (These tools are widely
used in the Python community and would make it easier for contributors to run
the tests.)

I had to fix two tests to be able to run them successfully.