Update caches instead of clearing them when calling removeChild()
This prevents cache corruptions when an element is removed and another
one is immediately added to a document. In such a situation, the element
cache of the document only contains the latter element. Since this cache
is not empty any more, further calls to other methods (such as
getElementsByType) do not trigger cache rebuilding, yielding to
incorrect results.